+++ program_management/ddd/service/write/delete_program_tree_version_service.py
@@ -23,82 +23,18 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from typing import List
-from django.db.models import Q
 
-from base.models.group_element_year import GroupElementYear
-from education_group.models.group_year import GroupYear
 from program_management.ddd.command import DeleteProgramTreeVersionCommand
-from program_management.models.education_group_version import EducationGroupVersion
-from program_management.models.element import Element
-from program_management.ddd.repositories.program_tree_version import ProgramTreeVersionRepository
 from program_management.ddd.domain.program_tree_version import ProgramTreeVersionIdentity
+from program_management.ddd.repositories.program_tree_version import ProgramTreeVersionRepository
 
 
-def delete_program_tree_version(command: DeleteProgramTreeVersionCommand) -> None:
-    education_group_versions = EducationGroupVersion.objects \
-        .select_related('root_group__element').filter(offer__acronym=command.offer_acronym,
-                                                      offer__academic_year__year__gte=command.year,
-                                                      version_name=command.version_name,
-                                                      is_transition=command.is_transition)
-    # TODO : la validation ne doit pas être ici???
-    # if DeleteVersionValidator(education_group_versions=education_group_versions).validate():
-    _delete_version_trees(education_group_versions)
+def delete(command: DeleteProgramTreeVersionCommand) -> None:
+    identity = ProgramTreeVersionIdentity(
+        command.offer_acronym,
+        command.year,
+        command.version_name,
+        command.is_transition
+    )
 
-
-def _delete_version_trees(education_group_versions: List[EducationGroupVersion]) -> None:
-    for education_group_version in education_group_versions:
-        identity_standard = ProgramTreeVersionIdentity(
-            education_group_version.offer.acronym,
-            education_group_version.offer.academic_year.year,
-            education_group_version.version_name,
-            education_group_version.is_transition
-        )
-
-        tree = ProgramTreeVersionRepository().get(entity_id=identity_standard)
-        _delete_version_tree_content(tree.get_tree())
-
-    for education_group_version in education_group_versions:
-        _delete_version_root_group(education_group_version)
-
-
-def _delete_version_root_group(education_group_version: EducationGroupVersion) -> None:
-    root_group_to_delete = education_group_version.root_group
-    education_group_version.delete()
-    if not Element.objects.filter(group_year=root_group_to_delete).exists():
-        root_group_to_delete.delete()
-
-
-def _delete_version_tree_content(tree: 'ProgramTree') -> None:
-    """
-    This function will delete group year and the default structure
-    """
-    group_year_ids = []
-    for node in tree.get_all_nodes():
-        #  TODO léger doute ici...
-        child_links_to_delete = GroupElementYear.objects.filter(
-            Q(parent_element__pk=node.pk) | Q(child_element__pk=node.pk)
-        )
-
-        elt_to_delete = Element.objects.filter(pk=node.pk).first()
-        if elt_to_delete:
-            group_year_ids.append(elt_to_delete.group_year.id)
-
-        for group_element_yr_to_delete in child_links_to_delete:
-            group_element_yr_to_delete.delete()
-        if elt_to_delete:
-            elt_to_delete.delete()
-
-    _delete_unused_group_year(group_year_ids)
-
-
-def _delete_unused_group_year(group_year_ids: List[int]) -> None:
-    for group_yr_id in group_year_ids:
-        if not GroupElementYear.objects.filter(child_element__group_year__pk=group_yr_id).exists() and \
-                not GroupElementYear.objects.filter(parent_element__group_year__pk=group_yr_id).exists() and \
-                not EducationGroupVersion.objects.filter(root_group__pk=group_yr_id).exists() and \
-                not Element.objects.filter(group_year__pk=group_yr_id).exists():
-            # No reuse
-            group_yr = GroupYear.objects.filter(pk=group_yr_id).first()
-            if group_yr:
-                group_yr.delete()
+    ProgramTreeVersionRepository().delete(entity_id=identity)
